Preoload the live container version in the settings and setup form children components.
