+++ assets/js/modules/analytics/components/settings/SettingsEdit.js
@@ -28,7 +28,6 @@ import useExistingTagEffect from '../../hooks/useExistingTagEffect';
 import useExistingGA4TagEffect from '../../../analytics-4/hooks/useExistingTagEffect';
 import SettingsForm from './SettingsForm';
 import { AccountCreate, AccountCreateLegacy } from '../common';
-import { MODULES_TAGMANAGER } from '../../../tagmanager/datastore/constants';
 const { useSelect } = Data;
 
 export default function SettingsEdit() {
@@ -61,21 +60,6 @@ export default function SettingsEdit() {
 			'analytics-4'
 		);
 	} );
-	const isTagManagerAvailable = useSelect( ( select ) =>
-		select( CORE_MODULES ).isModuleAvailable( 'tagmanager' )
-	);
-	// Preloading the `live-container-version` call is necessary to display
-	// the loading state until the data is fetched from the server.
-	// This call is being made in the child component ExistingGTMPropertyNotice
-	// to ensure that the loading state is displayed before the data is available.
-	useSelect(
-		( select ) =>
-			isTagManagerAvailable &&
-			select( MODULES_TAGMANAGER ).getSingleAnalyticsPropertyID()
-	);
-	const gtmContainersResolved = useSelect( ( select ) =>
-		select( MODULES_TAGMANAGER ).hasFinishedLoadingContainers()
-	);
 
 	useExistingTagEffect();
 	useExistingGA4TagEffect();
@@ -88,7 +72,6 @@ export default function SettingsEdit() {
 	if (
 		isDoingSubmitChanges ||
 		! hasResolvedAccounts ||
-		! gtmContainersResolved ||
 		hasAnalyticsAccess === undefined ||
 		hasAnalytics4Access === undefined
 	) {
